requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
certifi==2024.2.2
click==8.1.8
discord-ext-ipcx==0.4.0
discord.py==2.5.2
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
multidict==6.3.2
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
python-dotenv==1.1.0
setuptools==75.8.0
Werkzeug==3.1.3
yarl==1.18.3
